Add files via upload
Updated hose pulling controller. The controller now uses the force prediction model trained on the robot joint positions mapped to the force data. In order to embed this force prediction model into a controller which works in Cartesian space, it was necessary to learn a model of the inverse kinematics. This was done by first finding the forward kinematics and then using that data to learn an inverse kinematics model.